Adding LA-PDES benchmarks for ROSS engine
Adding an implementation of the LA-PDES benchmarks for the ROSS PDES engine. The archive file is distributed under the ROSS license and is included here in the hope that other researchers would find it useful. No claims as to completeness or efficiency in implementation is being made. Also no support for people who choose to use it.